Create Segregate 0s and 1s
Given an array of length N consisting of only 0s and 1s in random order. Modify the array to segregate 0s on left side and 1s on the right side of the array.
We solve this question using two pointer approach.
To iterate on array we use while loop with two variable s for starting index and e for ending index.