Create  Code Testcase Testcase Test Result 70. Climbing Stairs Solved(LEETCODE)
You are climbing a staircase. It takes n steps to reach the top.

Each time you can either climb 1 or 2 steps. In how many distinct ways can you climb to the top?

we use pull dp to update an array of values that contain the number of ways to reach the nth step. Pull DP is a type of dynamic programing where we solve the first tiny problem and then when we go to the next problem or in our case, the next step, we simply use the previously computed values.