532. K-diff Pairs in an Array(LEETCODE.cpp
Given an array of integers nums and an integer k, return the number of unique k-diff pairs in the array.

A k-diff pair is an integer pair (nums[i], nums[j]), where the following are true:

0 <= i, j < nums.length
i != j
|nums[i] - nums[j]| == k
Notice that |val| denotes the absolute value of val.

SOLUTION :
1. USING MAPS :
We are aware of the fact that for a pair to be counted as an answer , both the elements ( x and x+k ) , need to be in the array.
So we simply create a map and store the frequency of each element in the map.
Now we traverse the map and for each element 'x' , we check if 'x+k' exists in the map . If it does , then it means a unique pair can be formed and hence, we increment the answer.

EDGE CASE :
The only edge case is the situation where k=0. If k=0 , instead of finding 'x+k' , we check if the frequency of 'x'>1. If it is , then we increment the answer .
Else , we don't increment the answer , as the frequency of x=1 and hence it can't form a pair with itself.